eval/pyproject.toml
[project]
name = "balkanbench"
version = "0.1.0.dev0"
description = "Reproducible evaluation benchmark for BCMS language models."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "BalkanBench contributors" }]
keywords = ["benchmark", "nlp", "evaluation", "serbian", "superglue", "bcms"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "typer>=0.12",
    "pydantic>=2.7",
    "jsonschema>=4.22",
    "pyyaml>=6.0",
    "accelerate>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML",
    "types-jsonschema",
    # publish extras: in dev so tests can exercise the subcommand
    "datasets>=2.20",
    "huggingface_hub>=0.24",
    # ml extras: in dev so CI runs the eval pipeline tests end to end
    "torch>=2.1",
    "transformers>=4.40",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "optuna>=3.6",
]
publish = [
    "datasets>=2.20",
    "huggingface_hub>=0.24",
    "hf-transfer>=0.1",
    "google-cloud-storage>=2.10",
]
ml = [
    "torch>=2.1",
    "transformers>=4.40",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "optuna>=3.6",
]

[project.scripts]
balkanbench = "balkanbench.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/balkanbench"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B"]
# typer uses function calls in argument defaults as the canonical idiom; B008 would
# force awkward module-level sentinels for every Option/Argument.
"src/balkanbench/cli/**" = ["B008"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["balkanbench"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "jsonschema.*",
    "datasets",
    "datasets.*",
    "huggingface_hub",
    "huggingface_hub.*",
    "torch",
    "torch.*",
    "transformers",
    "transformers.*",
    "sklearn",
    "sklearn.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=balkanbench --cov-report=term-missing --cov-report=xml --cov-fail-under=80"

[tool.coverage.run]
source = ["src/balkanbench"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
